refactor(gallery): compile shared sources once
