+++ app/CMakeLists.txt
@@ -1,6 +1,6 @@
 include("${CMAKE_CURRENT_SOURCE_DIR}/GalleryMetadata.cmake")
 
-file(GLOB_RECURSE FLUENT_QT_GALLERY_SOURCES CONFIGURE_DEPENDS
+file(GLOB_RECURSE FLUENT_QT_GALLERY_CORE_SOURCES CONFIGURE_DEPENDS
     ${CMAKE_CURRENT_SOURCE_DIR}/model/*.h
     ${CMAKE_CURRENT_SOURCE_DIR}/model/*.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/view/*.h
@@ -17,44 +17,65 @@ set(FLUENT_QT_GALLERY_WEB_RESOURCES
 if(EMSCRIPTEN)
     list(APPEND FLUENT_QT_GALLERY_RESOURCES
         ${FLUENT_QT_GALLERY_WEB_RESOURCES})
-    list(FILTER FLUENT_QT_GALLERY_SOURCES EXCLUDE REGEX
+    list(FILTER FLUENT_QT_GALLERY_CORE_SOURCES EXCLUDE REGEX
         "/Gallery(ApplicationController|SingleInstance|WindowPlacement)\\.(cpp|h)$")
-    list(FILTER FLUENT_QT_GALLERY_SOURCES EXCLUDE REGEX
+    list(FILTER FLUENT_QT_GALLERY_CORE_SOURCES EXCLUDE REGEX
         "/GalleryCloseBehaviorPrompt\\.(cpp|h)$")
-    list(FILTER FLUENT_QT_GALLERY_SOURCES EXCLUDE REGEX
+    list(FILTER FLUENT_QT_GALLERY_CORE_SOURCES EXCLUDE REGEX
         "/view/preview/.*\\.(cpp|h)$")
 endif()
 source_group(TREE ${CMAKE_CURRENT_SOURCE_DIR} FILES
-    ${FLUENT_QT_GALLERY_SOURCES}
+    ${FLUENT_QT_GALLERY_CORE_SOURCES}
     ${FLUENT_QT_GALLERY_RESOURCES}
     "${CMAKE_CURRENT_SOURCE_DIR}/platform/GalleryPlatform.h"
 )
 
 set(FLUENT_QT_GALLERY_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR} PARENT_SCOPE)
-set(FLUENT_QT_GALLERY_SOURCES ${FLUENT_QT_GALLERY_SOURCES} PARENT_SCOPE)
 set(FLUENT_QT_GALLERY_RESOURCES ${FLUENT_QT_GALLERY_RESOURCES} PARENT_SCOPE)
 set(FLUENT_QT_GALLERY_WEB_RESOURCES ${FLUENT_QT_GALLERY_WEB_RESOURCES} PARENT_SCOPE)
 
 find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Network)
 
+# The application and Gallery test executables share the same model/view code.
+# Keep it in one internal object target so a cold full build compiles those sources
+# once; platform adapters and qrc files stay on their final executable targets.
+add_library(FluentQtGalleryCore OBJECT
+    ${FLUENT_QT_GALLERY_CORE_SOURCES}
+)
+fluent_qt_configure_cpp_target(FluentQtGalleryCore)
+fluent_qt_enable_sanitizers(FluentQtGalleryCore)
+target_include_directories(FluentQtGalleryCore
+    PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_link_libraries(FluentQtGalleryCore
+    PRIVATE
+    FluentQt::FluentQt
+    FluentQtLoggingSupport
+    Qt${QT_VERSION_MAJOR}::Widgets
+    Qt${QT_VERSION_MAJOR}::Network
+)
+# Consumers already link these runtime dependencies. The object target needs
+# their compile usage requirements only; suppress link-only propagation so
+# Apple ld does not receive duplicate FluentQt/logging archives.
+set_property(TARGET FluentQtGalleryCore PROPERTY INTERFACE_LINK_LIBRARIES "")
+
 if(COMMAND qt_add_executable)
     qt_add_executable(${FLUENT_QT_GALLERY_EXECUTABLE_NAME}
         main.cpp
-        ${FLUENT_QT_GALLERY_SOURCES}
         ${FLUENT_QT_GALLERY_RESOURCES}
     )
 else()
     add_executable(${FLUENT_QT_GALLERY_EXECUTABLE_NAME}
         main.cpp
-        ${FLUENT_QT_GALLERY_SOURCES}
         ${FLUENT_QT_GALLERY_RESOURCES}
     )
 endif()
 fluent_qt_configure_cpp_target(${FLUENT_QT_GALLERY_EXECUTABLE_NAME})
+fluent_qt_enable_sanitizers(${FLUENT_QT_GALLERY_EXECUTABLE_NAME})
 target_sources(${FLUENT_QT_GALLERY_EXECUTABLE_NAME} PRIVATE
     "${CMAKE_CURRENT_SOURCE_DIR}/platform/GalleryPlatform.h")
 target_link_libraries(${FLUENT_QT_GALLERY_EXECUTABLE_NAME}
     PRIVATE
+    FluentQtGalleryCore
     FluentQt::FluentQt
     FluentQtLoggingSupport
     Qt${QT_VERSION_MAJOR}::Widgets
