+++ docs/development/production-evidence.md
@@ -38,6 +38,41 @@ set size was 40,812,544 bytes (38.9 MiB); median macOS peak memory footprint was
 12,142,016 bytes (11.6 MiB). The same generated project passed its application
 test, offscreen UI smoke, and Inspector quality report (zero findings).
 
+## Gallery cold-build baseline
+
+Recorded on 2026-08-28 on the same MacBook Air (Apple M2, 8 logical cores,
+24 GB) using macOS arm64, Qt 6.9.3, AppleClang 17, CMake 3.26.4, the
+`vcpkg-osx` Debug preset, and Unix Makefiles. The independent
+`build/benchmark-vcpkg-osx` directory did not reuse project object files;
+vcpkg dependencies were restored from the existing binary cache.
+
+Configuration completed in 5.07 seconds wall time. The cold build covered the
+shared Gallery core, the desktop Gallery, and both Gallery shell/content test
+executables:
+
+```bash
+/usr/bin/time -p python3 tools/dev/fluent_qt_build.py \
+  build/benchmark-vcpkg-osx \
+  --target FluentQtGalleryCore fluent_qt_gallery \
+           test_gallery_shell_framework test_gallery_content_pages
+```
+
+The adaptive wrapper selected five jobs from an eight-core CPU limit and a
+five-job memory limit (9.8 GiB currently available, 1.0 GiB reserved, 1.5 GiB
+per job). The cold build completed in 77.20 seconds wall time, with 298.53
+seconds user CPU and 33.67 seconds system CPU. An immediate no-change run
+completed in 4.36 seconds wall time; the wrapper selected six jobs after
+available memory rose to 10.0 GiB.
+
+The object-library boundary is the structural result behind this measurement:
+the 59 Gallery model/view/viewmodel translation units in the measured tree were
+compiled once instead of once for each of the application and two test
+executables, removing 118 duplicate compile actions for that source set. The
+subsequent TreeView sample split leaves 60 smaller translation units compiled
+once, rather than 180 compile actions under the former target layout. This is
+one local point-in-time baseline, not a universal build-time improvement claim;
+the checked build graph and tests remain the regression gates.
+
 ## DataGrid large-model baseline
 
 Recorded on 2026-08-25 from commit
